+++ database/migrations/2025_08_31_105856_create_room_session_reservations_table.php
@@ -10,13 +10,9 @@ return new class extends Migration {
     private const ID = "id";
     private const CREATED_AT = "created_at";
     private const UPDATED_AT = "updated_at";
-    private const DELETED_AT = "deleted_at";
     private const TABLE_NAME = "sipr_room_session_reservations";
-    
     private const TABLE_ROOM_RESERVATION = "sipr_room_reservations";
     private const TABLE_ROOM_RESERVATION_COLUMN_ID = "id";
-    private const TABLE_USER_COLUMN_ID = "id";
-    private const USER_ID = "user_id";
     private const TABLE_ROOM_SESSION = 'sipr_room_sessions';
     private const TABLE_ROOM_SESSION_COLUMN_ID = 'id';
     private const ROOM_RESERVATION_ID = "room_reservation_id";
@@ -27,43 +23,45 @@ return new class extends Migration {
      */
     public function up(): void
     {
-        Schema::create(self::TABLE_NAME, function (Blueprint $table) {
-            $table
-                ->id(self::ID)
-                ->autoIncrement()
-                ->unsigned()
-                ->nullable(false);
+        if (Schema::hasTable(self::TABLE_NAME)) {
+            # code...
+        } else {
+            Schema::create(self::TABLE_NAME, function (Blueprint $table) {
+                $table
+                    ->id(self::ID)
+                    ->autoIncrement()
+                    ->unsigned()
+                    ->nullable(false);
 
-            $table->unsignedBigInteger(self::ROOM_SESSION_ID)->nullable();
-            $table->unsignedBigInteger(self::ROOM_RESERVATION_ID)->nullable();
+                $table->unsignedBigInteger(self::ROOM_SESSION_ID)->nullable();
+                $table->unsignedBigInteger(self::ROOM_RESERVATION_ID)->nullable();
 
-            $table
-                ->foreign(self::ROOM_RESERVATION_ID)
-                ->references(self::TABLE_ROOM_RESERVATION_COLUMN_ID)
-                ->on(self::TABLE_ROOM_RESERVATION)
-                ->cascadeOnUpdate()
-                ->cascadeOnDelete();
+                $table
+                    ->foreign(self::ROOM_RESERVATION_ID)
+                    ->references(self::TABLE_ROOM_RESERVATION_COLUMN_ID)
+                    ->on(self::TABLE_ROOM_RESERVATION)
+                    ->cascadeOnUpdate()
+                    ->cascadeOnDelete();
 
-            $table
-                ->foreign(self::ROOM_SESSION_ID)
-                ->references(self::TABLE_ROOM_SESSION_COLUMN_ID)
-                ->on(self::TABLE_ROOM_SESSION)
-                ->cascadeOnUpdate()
-                ->cascadeOnDelete();
+                $table
+                    ->foreign(self::ROOM_SESSION_ID)
+                    ->references(self::TABLE_ROOM_SESSION_COLUMN_ID)
+                    ->on(self::TABLE_ROOM_SESSION)
+                    ->cascadeOnUpdate()
+                    ->cascadeOnDelete();
 
-            $table
-                ->dateTime(self::CREATED_AT)
-                ->nullable(true)
-                ->useCurrent();
+                $table
+                    ->dateTime(self::CREATED_AT)
+                    ->nullable(true)
+                    ->useCurrent();
 
-            $table
-                ->dateTime(self::UPDATED_AT)
-                ->nullable(true)
-                ->useCurrent()
-                ->useCurrentOnUpdate();
-        });
-
-        (new \Database\Seeders\DatabaseSeeder)->run();
+                $table
+                    ->dateTime(self::UPDATED_AT)
+                    ->nullable(true)
+                    ->useCurrent()
+                    ->useCurrentOnUpdate();
+            });
+        }
     }
 
     /**
@@ -71,6 +69,12 @@ return new class extends Migration {
      */
     public function down(): void
     {
-        Schema::dropIfExists(self::TABLE_NAME);
+        if (Schema::hasTable(self::TABLE_NAME)) {
+            // Jika ada, lakukan drop
+            Schema::dropIfExists(self::TABLE_NAME);
+
+            // Anda bisa menambahkan log atau pesan di sini jika diperlukan
+            // Log::info("Tabel " . self::TABLE_NAME . " berhasil dihapus.");
+        }
     }
 };
